+++ src/main/java/com/starkbank/BrcodePreview.java
@@ -18,8 +18,6 @@ public final class BrcodePreview extends SubResource {
     * name [string]: Payment receiver name. ex: "Tony Stark"
     * taxId [string]: Payment receiver tax ID. ex: "012.345.678-90"
     * bankCode [string]: Payment receiver bank code. ex: "20018183"
-    * branchCode [string]: Payment receiver branch code. ex: "0001"
-    * accountNumber [string]: Payment receiver account number. ex: "1234567"
     * accountType [string]: Payment receiver account type. ex: "checking"
     * allowChange [bool]: If True, the payment is able to receive amounts that are different from the nominal one. ex: True or False
     * amount [integer]: Value in cents that this payment is expecting to receive. If 0, any value is accepted. ex: 123 (= R$1,23)
@@ -36,8 +34,6 @@ public final class BrcodePreview extends SubResource {
     public String name;
     public String taxId;
     public String bankCode;
-    public String branchCode;
-    public String accountNumber;
     public String accountType;
     public Boolean allowChange;
     public long amount;
@@ -58,8 +54,6 @@ public final class BrcodePreview extends SubResource {
     * @param name [string]: Payment receiver name. ex: "Tony Stark"
     * @param taxId [string]: Payment receiver tax ID. ex: "012.345.678-90"
     * @param bankCode [string]: Payment receiver bank code. ex: "20018183"
-    * @param branchCode [string]: Payment receiver branch code. ex: "0001"
-    * @param accountNumber [string]: Payment receiver account number. ex: "1234567"
     * @param accountType [string]: Payment receiver account type. ex: "checking"
     * @param allowChange [bool]: If True, the payment is able to receive amounts that are different from the nominal one. ex: True or False
     * @param amount [integer]: Value in cents that this payment is expecting to receive. If 0, any value is accepted. ex: 123 (= R$1,23)
@@ -70,16 +64,14 @@ public final class BrcodePreview extends SubResource {
     * @param discountAmount [integer]: Current discount value in cents that this payment is expecting. ex: 123 (= R$1,23)
     * @param reconciliationId [string]: Reconciliation ID linked to this payment. ex: "txId", "payment-123"
     */
-    public BrcodePreview(String status, String name, String taxId, String bankCode, String branchCode, String accountNumber,
-                         String accountType, boolean allowChange, long amount, long nominalAmount, long interestAmount,
-                         long fineAmount, long reductionAmount, long discountAmount, String reconciliationId
+    public BrcodePreview(String status, String name, String taxId, String bankCode, String accountType,
+                         boolean allowChange, long amount, long nominalAmount, long interestAmount,
+                         long fineAmount,  long reductionAmount, long discountAmount, String reconciliationId
     ) {
         this.status = status;
         this.name = name;
         this.taxId = taxId;
         this.bankCode = bankCode;
-        this.branchCode = branchCode;
-        this.accountNumber = accountNumber;
         this.accountType = accountType;
         this.allowChange = allowChange;
         this.amount = amount;
@@ -102,8 +94,6 @@ public final class BrcodePreview extends SubResource {
     * name [string]: Payment receiver name. ex: "Tony Stark"
     * taxId [string]: Payment receiver tax ID. ex: "012.345.678-90"
     * bankCode [string]: Payment receiver bank code. ex: "20018183"
-    * branchCode [string]: Payment receiver branch code. ex: "0001"
-    * accountNumber [string]: Payment receiver account number. ex: "1234567"
     * accountType [string]: Payment receiver account type. ex: "checking"
     * allowChange [bool]: If True, the payment is able to receive amounts that are different from the nominal one. ex: True or False
     * amount [integer]: Value in cents that this payment is expecting to receive. If 0, any value is accepted. ex: 123 (= R$1,23)
@@ -122,8 +112,6 @@ public final class BrcodePreview extends SubResource {
         this.name = (String) dataCopy.remove("name");
         this.taxId = (String) dataCopy.remove("taxId");
         this.bankCode = (String) dataCopy.remove("bankCode");
-        this.branchCode = (String) dataCopy.remove("branchCode");
-        this.accountNumber = (String) dataCopy.remove("accountNumber");
         this.accountType = (String) dataCopy.remove("accountType");
         this.allowChange = (boolean) dataCopy.remove("allowChange");
         this.amount = (long) dataCopy.remove("amount");